TASK-54447: Split less styles for tasks comments drawer for better reuse and integration outside of the tasks app
plit less styles for tasks comments drawer for better reuse and integration outside of the tasks app
