+++ task-management/src/main/webapp/skin/css/tasks.less
@@ -2756,268 +2756,6 @@
       display: block;
     }
   }
-  .taskCommentsAndChanges {
-    .taskComments {
-      .addCommentBtn {
-        .uiIconTaskAddComment {
-          position: relative;
-          cursor: pointer;
-          margin-top: 2px;
-          &:before {
-            content: '\e620';
-            font-family: "PLF-FONT-ICONS";
-            font-size: 16px;
-            color: @primaryColor;
-          }
-          &:hover {
-            &:before {
-              color: @primaryColor
-            }
-          }
-        }
-      }
-      .lastCommentLabel {
-        font-size: 16px;
-        color: @greyColorLighten1;
-      }
-      .ViewAllCommentLabel {
-        font-size: 16px;
-        color: @primaryColor;
-        text-decoration: underline;
-        cursor: pointer;
-      }
-      .comment{
-        position: relative;
-      }
-      .commentItem {
-        width: 100%;
-        .commentHeader {
-          width: 100%;
-          .removeCommentBtn {
-            margin-left: auto ~'; /** orientation=lt */ ';
-            margin-right: auto ~'; /** orientation=rt */ ';
-            margin-top: -4px;
-          }
-          .externalTagClass {
-            color: @baseColorLight;
-            font-weight: normal;
-          }
-        }
-        .commentBody {
-          display: inline-block;
-          .taskContentComment {
-            background: @commentBackgroundColor;
-            padding: 10px;
-            border-radius: 10px;
-            word-break: break-word;
-            p {
-              margin-bottom: 0!important;
-            }
-            .externalTagClass {
-              color: @baseColorLight;
-              font-weight: normal;
-            }
-          }
-        }
-      }
-      .TaskSubComments{
-        margin-left: -20px ~'!important; /** orientation=lt */ ';
-        margin-right: -20px ~'!important; /** orientation=rt */ ';
-        .TaskSubCommentItem {
-          padding-left: 60px ~'; /** orientation=lt */ ';
-          padding-right: 60px ~'; /** orientation=rt */ ';
-        }
-      }
-      .taskCommentEmpty {
-        .uiIconComment{
-          display: block;
-          margin-bottom: 10px;
-          &:before {
-            font-family: "PLF-FONT-ICONS" !important;
-            content: '\e9c1';
-            font-style: normal;
-            font-size: 34px;
-            color: @greyColor;
-          }
-        }
-        .noCommentLabel {
-          display: block;
-          color: @greyColorLighten1;
-          font-size: 16px;
-        }
-        .ViewAllCommentText {
-          display: block;
-          text-decoration: underline;
-          color: @primaryColor;
-          font-size: 14px;
-          cursor: pointer;
-        }
-      }
-    }
-    .commentBtn {
-      background-color: @primaryColor!important;
-      letter-spacing: 0!important;
-      font-size: 14px!important;
-      text-transform: capitalize;
-    }
-  }
-}
-
-.taskCommentDrawer {
-  .addCommentBtn {
-    .uiIconTaskAddComment {
-      position: relative;
-      cursor: pointer;
-      top: 0px;
-      &:before {
-        content: '\e620';
-        font-family: "PLF-FONT-ICONS";
-        font-size: 18px;
-        color: @primaryColor;
-      }
-      &:hover {
-        &:before {
-          color: @primaryColor
-        }
-      }
-    }
-  }
-  .drawerTitle > * {
-    flex: 0 0 auto!important;
-  }
-  .uiArrowBAckIcon {
-    cursor: pointer;
-    color: @btnDisableColor!important;
-    margin-left: -4px ~'; /** orientation=lt */ ';
-    margin-right: -4px ~'; /** orientation=rt */ ';
-    &:before {
-      content: "\e913";
-      font-family: "PLF-FONT-ICONS" !important;
-      font-style: normal;
-      font-size: 22px;
-    }
-  }
-  .drawerContent {
-    padding: 8px;
-    .dateTime.caption {
-      color: @greyColorLighten1;
-      text-align: start;
-      padding-left: 10px ~'; /** orientation=lt */ ';
-      padding-right: 10px ~'; /** orientation=rt */ ';
-    }
-    .TaskCommentContent {
-      .commentItem {
-        width: 100%;
-        .commentHeader {
-          width: 100%;
-          .removeCommentBtn {
-            margin-left: auto ~'; /** orientation=lt */ ';
-            margin-right: auto ~'; /** orientation=rt */ ';
-            .uiIconTrashMini {
-              font-size: 12px;
-            }
-          }
-          .externalTagClass {
-            color: @baseColorLight;
-            font-weight: normal;
-          }
-        }
-        .commentBody {
-          display: inline-block;
-          .taskContentComment {
-            background: @commentBackgroundColor;
-            padding: 10px;
-            border-radius: 10px;
-            word-break: break-word;
-            p {
-              margin-bottom: 0!important;
-            }
-            .externalTagClass {
-              color: @baseColorLight;
-              font-weight: normal;
-            }
-          }
-        }    
-      }
-      .TaskSubComments{
-        margin-left: -20px ~'!important; /** orientation=lt */ ';
-        margin-right: -20px ~'!important; /** orientation=rt */ ';
-        .TaskSubCommentItem {
-          padding-left: 60px ~'; /** orientation=lt */ ';
-          padding-right: 60px ~'; /** orientation=rt */ ';
-        }
-      }
-      .commentEditorContainer {
-        padding-left: 40px ~'; /** orientation=lt */ ';
-        padding-right: 40px ~'; /** orientation=rt */ ';
-      }
-      .newCommentEditor {
-        padding-left: 0 ~'; /** orientation=lt */ ';
-        padding-right: 0 ~'; /** orientation=rt */ ';
-      }
-    }
-    .editorContent {
-      .inactiveEditor {
-        display: none !important;
-      }
-      .comment {
-        position: relative;
-        .editorContainer {
-          width: 100%;
-        }
-      }
-      .tooManyChars {
-        color: #bc4343!important;
-      }
-      .tooManyChars .uiIconMessageLength {
-        color: #bc4343!important;
-      }
-      .activityCharsCount{
-        position: absolute;
-        bottom: 46px;
-        font-size: 12px;
-        right: 3px ~'; /** orientation=lt */ ';
-        left: 3px ~'; /** orientation=rt */ ';
-        .uiIconMessageLength {
-          font-size: 14px;
-          color: #2eb58c;
-          &:before {
-            content: "\e910";
-          }
-        }
-      }
-      .cke_bottom.cke_reset_all {
-        background-color: transparent;
-        height: 30px!important;
-      }
-      .cke_toolbar {
-        border-right: none ~'!important; /** orientation=lt */ ';
-        border-left: none ~'!important; /** orientation=rt */ ';
-      }
-      .cke_bottom .cke_button_icon {
-        filter: opacity(0.4);
-      }
-      .tooManyChars {
-        color: #bc4343!important;
-      }
-      .tooManyChars .uiIconMessageLength {
-        color: #bc4343!important;
-      }
-    }
-    .changeUserAvatar {
-      .pull-left.ml-2.d-flex.flex-column {
-        align-self: flex-start;
-        margin-top: -5px;
-        p {
-          font-weight: bold;
-          color: @primaryColor !important;
-        }
-      }
-    }
-    .changesText {
-      max-width: 270px;
-    }
-  }
 }
 
 @-webkit-keyframes slideInRight {
